requirements.txt
Django
djangorestframework
djoser
django-cors-headers
psycopg2-binary
